seperate testrunner from harness.py
